Chapter 13 Rename to just camera_stream.
